backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
sqlparse
psycopg2-binary
python-dotenv
mysqlclient
pillow